AI_mini/backend/requirements.txt
fastapi
uvicorn
google-genai
python-multipart
sqlalchemy
pydantic
python-dotenv
